updated "uses" in each header, moved enums to a separate package outside of the model package
