+++ DanceGame/core/src/com/mygdx/game/Enums/Color.java
@@ -1,4 +1,4 @@
-package com.mygdx.game.model.Enums;
+package com.mygdx.game.Enums;
 
 /**
  * A enum used to keep track of which color dancers have / which player dancers belong to. As of now we have red and
